Fix compiler warning (STR #1246)
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@4579 7a7537e8-13f0-0310-91df-b6672ffda945

+++ filter/image-png.c
@@ -83,7 +83,7 @@ ImageReadPNG(image_t    *img,		/* IO - Image */
   png_read_info(pp, info);
 
   fprintf(stderr, "DEBUG: PNG image: %dx%dx%d, color_type=%x (%s%s%s)\n",
-          info->width, info->height, info->bit_depth, info->color_type,
+          (int)info->width, (int)info->height, info->bit_depth, info->color_type,
 	  (info->color_type & PNG_COLOR_MASK_COLOR) ? "RGB" : "GRAYSCALE",
 	  (info->color_type & PNG_COLOR_MASK_ALPHA) ? "+ALPHA" : "",
 	  (info->color_type & PNG_COLOR_MASK_PALETTE) ? "+PALETTE" : "");
